+++ tests/archetype-quickstart/run.sh
@@ -12,6 +12,7 @@ mvn archetype:generate \
     -B \
     -DarchetypeGroupId=com.dimajix.flowman.maven \
     -DarchetypeArtifactId=flowman-archetype-quickstart \
+    -DarchetypeVersion=0.2.1 \
     -DgroupId=test \
     -DartifactId=quickstart-test \
     -Dversion=1.0-SNAPSHOT \